% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{add_plot}
\alias{add_plot}
\title{Add a plot to a DiseaseDash report}
\usage{
add_plot(
  report,
  outcome,
  colors = c(low = "#1aab4a", mid = "#e02f0b", high = "#d837ed")
)
}
\arguments{
\item{report}{The DiseaseDash report to add a plot to}

\item{outcome}{The outcome to be displayed in the plot. Either "cases" or "deaths"}

\item{colors}{The colors corresponding to the low, medium, and high points on the plot. Defaults are green, red, and violet shades.}
}
\value{
The DiseaseDash report object with a plot added
}
\description{
This function generates a map filled with a color gradient showing the change
in COVID-19 cases or deaths during the time interval specified in the report.
Currently, vaccination data isn't supported

Extensive customizations to this plot can be made using the `modify_plot` function
}
\examples{
\dontrun{
  my_report <- disease_dash(type = "cases",
                           region = "california",
                           start =  as.Date("2021-04-01"),
                           end = as.Date("2021-05-01")) \%>\%
              add_plot("cases")
}
}
