% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{add_table}
\alias{add_table}
\title{Add a table to a DiseaseDash report}
\usage{
add_table(report, outcome, rows = 5, order = "desc")
}
\arguments{
\item{report}{The report you are adding the table to}

\item{outcome}{Whether cases or deaths are being displayed. Should be either "cases" or "deaths"}

\item{rows}{How many rows the table should contain. Default is 5.}

\item{order}{Whether the rows should be sorted in ascending or descending order. Should be either "asc" or "desc"}
}
\value{
The report with a table added to it
}
\description{
Adds a table from **{flextable}** showing cases or deaths in the region
specified in the report. Note that the **{magick}** package is required to be
installed to render tables in a report
}
\examples{
\dontrun{
  my_report <- disease_dash(type = "cases",
                           region = "california",
                           start =  as.Date("2021-04-01"),
                           end = as.Date("2021-05-01")) \%>\%
              add_table("cases")
}
}
