% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_report}
\alias{build_report}
\title{Compile a DiseaseDash report}
\usage{
build_report(report, new_title = NULL)
}
\arguments{
\item{report}{The report to be built}

\item{new_title}{Optional. A new title to be displayed for the report. Default is "DiseaseDash report for region"}
}
\value{
A `grob` from the **{grid}** package that can be saved using ggplot2::ggsave()
}
\description{
Compile a DiseaseDash report
}
\examples{
\dontrun{
  library(ggplot2)
  library(dplyr)

  my_report <- report(type = "cases",
                      region = "california",
                      start = as.Date("2021-11-01"),
                      end = as.Date("2022-01-01")) \%>\%
              add_plot("cases") \%>\%
              add_table("death") \%>\%
              build_report()

 ggsave("my_report.pdf", my_report)
}
}
