% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cache.R
\name{check_cache}
\alias{check_cache}
\title{Check if a file exists in the DiseaseDash cache}
\usage{
check_cache(data_type)
}
\arguments{
\item{data_type}{The type of data to check for. Should be either "cases" or "vax"}
}
\value{
TRUE if the data type is stored in the cache, FALSE otherwise
}
\description{
Check if a file exists in the DiseaseDash cache
}
\examples{
\dontrun{
  if (check_cache("vax")) {
     # Do something here
   }
}
}
