% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{disease_dash}
\alias{disease_dash}
\title{Initialize a DiseaseDash report}
\usage{
disease_dash(type, region, load_cache = T, save_cache = T)
}
\arguments{
\item{type}{The type of report to make. Either "cases" or "vax"}

\item{region}{The region you want to show the data for. Either a U.S state or "national" for the whole U.S}

\item{load_cache}{Whether or not to load data from the cache. Recommended if you are running multiple reports.
Default is TRUE.}

\item{save_cache}{Whether or not to save data from the cache if you are downloading from a data source.
Default is TRUE.}
}
\value{
A DiseaseDash report object that can be piped through to other report functions
}
\description{
Initialize a DiseaseDash report
}
