% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasedash-package.R
\docType{package}
\name{DiseaseDash-package}
\alias{DiseaseDash}
\alias{DiseaseDash-package}
\title{DiseaseDash: Create reports on COVID-19 data}
\description{
This package utilizes data from the NYT COVID-19 tracking project to create informative dashboards at the national and state level.
}
