% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cache.R
\name{get_cache_info}
\alias{get_cache_info}
\title{Get DiseaseDash cache info file}
\usage{
get_cache_info()
}
\value{
A data.frame containing the data types stored in the cache and the
    last time they were updated. NULL if there is no cache
}
\description{
Reads the DiseaseDash cache info file if it exists. Returns NULL otherwise
}
\examples{
cache_info <- get_cache_info()
}
