% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\name{get_infection_data}
\alias{get_infection_data}
\title{Pull COVID-19 infection data}
\usage{
get_infection_data(cache = T, data_url = NULL)
}
\arguments{
\item{cache}{Should the data be cached to speed up times for making multiple reports? Default is true.
See details for more info}

\item{data_url}{Optional. An alternate URL pointing to the NYT COVID-19 data}
}
\value{
A data.frame of the latest NYT county-level COVID-19 data
}
\description{
This function downloads COVID-19 data from the NYT off of their Github repo
located at: https://github.com/nytimes/covid-19-data. In case the location has
changed, you can specify an updated URL using the data_url parameter. Most users
won't need this function, but it is here if you want to look at the data yourself
}
\examples{
\dontrun{covid_dat = get_infection_data()}
}
