% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify_data}
\alias{modify_data}
\title{Modify the data in a DiseaseDash report}
\usage{
modify_data(report, fn, ...)
}
\arguments{
\item{report}{The report you want to modify}

\item{fn}{The function you would like to apply to the data.}

\item{...}{Additional arguments for `fn` if needed}
}
\value{
The report passed to this function with modified data
}
\description{
Modify the data in a DiseaseDash report
}
\examples{
\dontrun{
  library(dplyr)

  my_report <- report(type = "cases",
                      region = "california",
                      start = as.Date("2021-11-01"),
                      end = as.Date("2022-01-01")) \%>\%
              modify_data(~filter(.x, county \%in\% c("alameda", "san francisco")))
}
}
