% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify_plot}
\alias{modify_plot}
\title{Modify the plot in a DiseaseDash report}
\usage{
modify_plot(report, fn, ...)
}
\arguments{
\item{report}{The report you want to modify}

\item{fn}{The function you would like to apply to the plot. Supports **{purrr}** style formulas}

\item{...}{Additional arguments for `fn` if needed}
}
\value{
The report passed to this data with a modified plot
}
\description{
Modify the plot in a DiseaseDash report
}
