% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cache.R
\name{read_cache}
\alias{read_cache}
\title{Read a data type from the DiseaseDash cache}
\usage{
read_cache(data_type)
}
\arguments{
\item{data_type}{The type of data you are retrieving. Either "cases" or "vax".}
}
\value{
A data.frame containing data on either COVID-19 cases or vaccinations
}
\description{
Read a data type from the DiseaseDash cache
}
\examples{
\dontrun{cases_data <- read_cache("cases")}
}
