% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_time.R
\name{specify_time}
\alias{specify_time}
\title{Specify the time intervals for a DiseaseDash report}
\usage{
specify_time(report, start, end, include_bounds = T)
}
\arguments{
\item{report}{The report you are specifying the time intervals for}

\item{start}{The start date of the time interval to include in the report}

\item{end}{The end date that should be included in the report}

\item{include_bounds}{Whether start and end should be included in the dates. Default is TRUE}
}
\value{
A DiseaseDash report object with the time interval specified
}
\description{
Specify the time intervals for a DiseaseDash report
}
