% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cache.R
\name{write_cache}
\alias{write_cache}
\title{Write to the DiseaseDash cache}
\usage{
write_cache(data, data_type)
}
\arguments{
\item{data}{The data to write to the cache}

\item{data_type}{The type of data you are saving. Either "cases" or "vax". Used to update the log}
}
\value{
No return
}
\description{
Write to the DiseaseDash cache
}
\examples{
cases_data <- get_infection_data()

write_cache(cases_data, "cases")
}
