% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify_data}
\alias{modify_data}
\title{Modify the data in a DiseaseDash report}
\usage{
modify_data(report, fn, ...)
}
\arguments{
\item{report}{The report you want to modify}

\item{fn}{The function you would like to apply to the data. Supports **{purrr}** style formulas}

\item{...}{Additional arguments for `fn` if needed}
}
\value{
The report passed to this function with modified data
}
\description{
Modify the data in a DiseaseDash report
}
