% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify_plot}
\alias{modify_plot}
\title{Modify the plot in a DiseaseDash report}
\usage{
modify_plot(report, fn, ...)
}
\arguments{
\item{report}{The report you want to modify}

\item{fn}{The function you would like to apply to the plot.}

\item{...}{Additional arguments for `fn` if needed}
}
\value{
The report passed to this data with a modified plot
}
\description{
Modify the plot in a DiseaseDash report
}
\examples{
\dontrun{
  library(dplyr)

  my_report <- report(type = "cases",
                      region = "california",
                      start = as.Date("2021-11-01"),
                      end = as.Date("2022-01-01")) \%>\%
              add_plot("cases") \%>\%
              modify_plot(function (x) x + ggplot2::labs(title = "Cases in California"))
}
}
